flask_server/requirements.txt
blinker==1.8.2
click==8.1.7
Flask==3.0.3
importlib_metadata==8.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
pillow==10.4.0
Werkzeug==3.0.3
zipp==3.19.2
